GH-2756 ShaclSail + RDFS + NativeStore sometimes deadlocks (#2757)
GH-2756 remove the shapes cache and simplify locking
