+++ core/sail/inferencer/src/main/java/org/eclipse/rdf4j/sail/inferencer/fc/SchemaCachingRDFSInferencerConnection.java
@@ -754,8 +754,10 @@ public class SchemaCachingRDFSInferencerConnection extends InferencerConnectionW
 
 		super.rollback();
 
-		sail.clearInferenceTables();
-		regenerateCacheAndInferenceMaps(false);
+		if (statementsAdded || statementsRemoved || schemaChange) {
+			sail.clearInferenceTables();
+			regenerateCacheAndInferenceMaps(false);
+		}
 
 		statementsRemoved = false;
 		statementsAdded = false;
