+++ core/sail/shacl/src/test/java/org/eclipse/rdf4j/sail/shacl/MultithreadedMemoryStoreTest.java
@@ -1,3 +1,10 @@
+/*******************************************************************************
+ * Copyright (c) 2019 Eclipse RDF4J contributors.
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Distribution License v1.0
+ * which accompanies this distribution, and is available at
+ * http://www.eclipse.org/org/documents/edl-v10.php.
+ *******************************************************************************/
 package org.eclipse.rdf4j.sail.shacl;
 
 import org.eclipse.rdf4j.sail.NotifyingSail;
